+++ pdfio-stream.c
@@ -616,7 +616,7 @@ _pdfioStreamOpen(pdfio_obj_t *obj,	// I - Object
     else
     {
       // Something else we don't support
-      _pdfioFileError(st->pdf, "Unsupported stream filter '/%s'.", filter);
+      _pdfioFileError(st->pdf, "Unsupported stream filter '%N'.", filter);
       goto error;
     }
   }
@@ -688,6 +688,10 @@ pdfioStreamPeek(pdfio_stream_t *st,	// I - Stream
 
 //
 // 'pdfioStreamPrintf()' - Write a formatted string to a stream.
+//
+// This function writes a formatted string to a stream.  In addition to the
+// standard `printf` format characters, you can use "%N" to format a PDF name
+// value ("/Name") and "%S" to format a PDF string ("(String)") value.
 //
 
 bool					// O - `true` on success, `false` on failure